Better title checking in PageTitle.
The PageTitle function assumed all URLs pointed to an HTML document. Because of
this, if it was not an HTML document a panic occured when trying to access the
second index of the slice returned by re.FindSubMatch. This fix attempts to
check that the URL is pointing to an HTML document by first sending a HEAD
request and checking the response headers. If the Content-Type header value
contains "text/html" then the function continues to check for a page title,
otherwise the URL itself is used at the title.